Changing infrastructure and adding the dynamic library.
