+++ XAnimatedImage.xcodeproj/xcuserdata/ple.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>SchemeUserState</key>
+	<dict>
+		<key>XAnimatedImage.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>0</integer>
+		</dict>
+		<key>XAnimatedImageDemo.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>1</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>7ED4070B1C51543100EBA64B</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>7ED4072D1C5154B800EBA64B</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>7ED407361C5154B800EBA64B</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>EA5BB0941C050CD000ED2FF5</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>EA5BB0A81C050CD000ED2FF5</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>EA5BB0B31C050CD000ED2FF5</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
+</dict>
+</plist>
