+++ XAnimatedImage/XAnimatedImage.h
@@ -1,0 +1,19 @@
+//
+//  XAnimatedImage.h
+//  XAnimatedImage
+//
+//  Created by Le, Phong on 1/21/16.
+//  Copyright © 2016 Khaled Taha. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for XAnimatedImage.
+FOUNDATION_EXPORT double XAnimatedImageVersionNumber;
+
+//! Project version string for XAnimatedImage.
+FOUNDATION_EXPORT const unsigned char XAnimatedImageVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <XAnimatedImage/PublicHeader.h>
+
+
